requirements.txt
seaborn
altair==5.0.1
jupyterlab==4.0.1
ipywidgets==8.0.6
jupyterlab-dash==0.1.0a3
python-dotenv==1.0.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.15
missingno==0.5.2
scikit-learn==1.2.2
